requirements.txt
numpy
pandas
scikit-learn
nltk
sacrebleu
transformers
openai
langchain
langchain-openai
langchain-community
langchain_core
deepeval==1.3.9
coverage
loguru
evaluate
rouge_score
bnlp_toolkit
